+++ tensorlayer/layers/object_detection.py
@@ -1,15 +1,15 @@
+#! /usr/bin/python
 # -*- coding: utf-8 -*-
 
-from tensorlayer import tl_logging as logging
 from tensorlayer.layers.core import Layer
+
+from tensorlayer import tl_logging as logging
 
 from tensorlayer.decorators import deprecated_alias
 
-try:
-    from tensorlayer.third_party.roi_pooling.roi_pooling.roi_pooling_ops import roi_pooling
-except Exception as e:
-    logging.error(e)
-    logging.error("HINT: 1. https://github.com/deepsense-ai/roi-pooling  2. tensorlayer/third_party/roi_pooling")
+from tensorlayer.lazy_imports import LazyImport
+
+roi_pooling = LazyImport("tensorlayer.third_party.roi_pooling.roi_pooling.roi_pooling_ops")
 
 __all__ = [
     'ROIPoolingLayer',
@@ -53,8 +53,6 @@ class ROIPoolingLayer(Layer):
 
         logging.info("ROIPoolingLayer %s: (%d, %d)" % (name, pool_height, pool_width))
 
-        self.inputs = prev_layer.outputs
-
         self.outputs = roi_pooling(self.inputs, rois, pool_height, pool_width)
 
-        self.all_layers.append(self.outputs)
+        self._add_layers(self.outputs)